+++ app/build.gradle.kts
@@ -50,4 +50,9 @@ dependencies {
     androidTestImplementation("androidx.test.espresso:espresso-core:3.5.1")
 
 
+    implementation ("com.squareup:javapoet:1.13.0")
+    implementation ("androidx.room:room-runtime:2.4.0") 
+    annotationProcessor ("androidx.room:room-compiler:2.4.0")
+
+
 }
