+++ app/src/main/AndroidManifest.xml
@@ -13,6 +13,12 @@
         android:theme="@style/Theme.Hotelbooking"
         tools:targetApi="31">
         <activity
+            android:name=".SettingNotificationActivity"
+            android:exported="false" />
+        <activity
+            android:name=".DetailNotificationActivity"
+            android:exported="false" />
+        <activity
             android:name=".MainActivity"
             android:exported="true"
             android:label="@string/app_name">
