+++ app/src/main/java/com/example/hotel_booking/entity/Notification.java
@@ -1,8 +1,10 @@
-package com.example.hotel_booking.model;
+package com.example.hotel_booking.entity;
 
 import org.threeten.bp.LocalDateTime;
 
-public class Notification {
+import java.io.Serializable;
+
+public class Notification implements Serializable {
     private int imgId;
     private String content;
