+++ app/src/main/java/com/example/hotel_booking/fragment/AccountViewModel.java
@@ -1,16 +1,16 @@
-package com.example.hotel_booking.ui.home;
+package com.example.hotel_booking.fragment;
 
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
 import androidx.lifecycle.ViewModel;
 
-public class HomeViewModel extends ViewModel {
+public class AccountViewModel extends ViewModel {
 
     private final MutableLiveData<String> mText;
 
-    public HomeViewModel() {
+    public AccountViewModel() {
         mText = new MutableLiveData<>();
-        mText.setValue("This is home fragment");
+        mText.setValue("This is account fragment");
     }
 
     public LiveData<String> getText() {
