+++ app/src/main/java/com/example/hotel_booking/fragment/HistoryFragment.java
@@ -1,4 +1,4 @@
-package com.example.hotel_booking.ui.account;
+package com.example.hotel_booking.fragment;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
@@ -10,22 +10,22 @@ import androidx.annotation.NonNull;
 import androidx.fragment.app.Fragment;
 import androidx.lifecycle.ViewModelProvider;
 
-import com.example.hotel_booking.databinding.FragmentAccountBinding;
+import com.example.hotel_booking.databinding.FragmentHistoryBinding;
 
-public class AccountFragment extends Fragment {
+public class HistoryFragment extends Fragment {
 
-    private FragmentAccountBinding binding;
+    private FragmentHistoryBinding binding;
 
     public View onCreateView(@NonNull LayoutInflater inflater,
                              ViewGroup container, Bundle savedInstanceState) {
-        AccountViewModel accountViewModel =
-                new ViewModelProvider(this).get(AccountViewModel.class);
+        HistoryViewModel historyViewModel =
+                new ViewModelProvider(this).get(HistoryViewModel.class);
 
-        binding = FragmentAccountBinding.inflate(inflater, container, false);
+        binding = FragmentHistoryBinding.inflate(inflater, container, false);
         View root = binding.getRoot();
 
-        final TextView textView = binding.textAccount;
-        accountViewModel.getText().observe(getViewLifecycleOwner(), textView::setText);
+        final TextView textView = binding.textHistory;
+        historyViewModel.getText().observe(getViewLifecycleOwner(), textView::setText);
         return root;
     }
