+++ app/src/main/java/com/example/hotel_booking/fragment/HistoryViewModel.java
@@ -1,16 +1,16 @@
-package com.example.hotel_booking.ui.notifications;
+package com.example.hotel_booking.fragment;
 
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
 import androidx.lifecycle.ViewModel;
 
-public class NotificationsViewModel extends ViewModel {
+public class HistoryViewModel extends ViewModel {
 
     private final MutableLiveData<String> mText;
 
-    public NotificationsViewModel() {
+    public HistoryViewModel() {
         mText = new MutableLiveData<>();
-        mText.setValue("This is notifications fragment");
+        mText.setValue("This is history fragment");
     }
 
     public LiveData<String> getText() {
