+++ app/src/main/java/com/example/hotel_booking/fragment/HomeViewModel.java
@@ -1,16 +1,16 @@
-package com.example.hotel_booking.ui.account;
+package com.example.hotel_booking.fragment;
 
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
 import androidx.lifecycle.ViewModel;
 
-public class AccountViewModel extends ViewModel {
+public class HomeViewModel extends ViewModel {
 
     private final MutableLiveData<String> mText;
 
-    public AccountViewModel() {
+    public HomeViewModel() {
         mText = new MutableLiveData<>();
-        mText.setValue("This is account fragment");
+        mText.setValue("This is home fragment");
     }
 
     public LiveData<String> getText() {
