--- app/src/main/java/com/example/hotel_booking/ui/history/HistoryViewModel.java
@@ -1,19 +1,0 @@
-package com.example.hotel_booking.ui.history;
-
-import androidx.lifecycle.LiveData;
-import androidx.lifecycle.MutableLiveData;
-import androidx.lifecycle.ViewModel;
-
-public class HistoryViewModel extends ViewModel {
-
-    private final MutableLiveData<String> mText;
-
-    public HistoryViewModel() {
-        mText = new MutableLiveData<>();
-        mText.setValue("This is history fragment");
-    }
-
-    public LiveData<String> getText() {
-        return mText;
-    }
-}
