+++ app/src/main/res/layout/activity_main.xml
@@ -13,7 +13,20 @@
         android:background="?attr/colorPrimary"
         app:layout_constraintTop_toTopOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"/>
+        app:layout_constraintEnd_toEndOf="parent">
+
+        <!-- Biểu tượng cài đặt -->
+        <ImageView
+            android:id="@+id/settings_icon"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_gravity="end"
+            android:src="@drawable/ic_settings"
+            android:paddingEnd="16dp"
+            android:contentDescription="@string/settings"
+            android:onClick="onSettingsClicked"/> <!-- Xử lý sự kiện khi người dùng nhấn -->
+
+    </androidx.appcompat.widget.Toolbar>
 
     <com.google.android.material.bottomnavigation.BottomNavigationView
         android:id="@+id/nav_view"
