+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -7,25 +7,25 @@
 
     <fragment
         android:id="@+id/navigation_home"
-        android:name="com.example.hotel_booking.ui.home.HomeFragment"
+        android:name="com.example.hotel_booking.fragment.HomeFragment"
         android:label="@string/title_home"
         tools:layout="@layout/fragment_home" />
 
     <fragment
         android:id="@+id/navigation_history"
-        android:name="com.example.hotel_booking.ui.history.HistoryFragment"
+        android:name="com.example.hotel_booking.fragment.HistoryFragment"
         android:label="@string/title_history"
         tools:layout="@layout/fragment_history" />
 
     <fragment
         android:id="@+id/navigation_notifications"
-        android:name="com.example.hotel_booking.ui.notifications.NotificationsFragment"
+        android:name="com.example.hotel_booking.fragment.NotificationsFragment"
         android:label="@string/title_notifications"
         tools:layout="@layout/fragment_notifications" />
 
     <fragment
         android:id="@+id/navigation_account"
-        android:name="com.example.hotel_booking.ui.account.AccountFragment"
+        android:name="com.example.hotel_booking.fragment.AccountFragment"
         android:label="@string/title_account"
         tools:layout="@layout/fragment_account" />
 </navigation>
