+++ app/src/main/res/values/colors.xml
@@ -3,4 +3,5 @@
     <color name="black">#FF000000</color>
     <color name="white">#FFFFFFFF</color>
     <color name="silver">#C0C0C0</color>
+    <color name="purple">#800080</color>
 </resources>
