+++ app/src/main/res/values/strings.xml
@@ -4,4 +4,5 @@
     <string name="title_history">History</string>
     <string name="title_notifications">Notifications</string>
     <string name="title_account">Account</string>
+    <string name="settings">Settings</string>
 </resources>
